Add table[column] namespacing helpers
If you have an ambiguous column, you can prefix it with a namespace by
simply subscripting the table with the column.

    let id = Expression<Int>("id")
    let manager_id = Expression<Int>("manager_id")
    let users = db["users"]
    let managers = users.alias("managers")

    users.join(managers, on: users[manager_id] == managers[id])

Unfortunately, Swift does not (yet) support subscript<T> (see
rdar://18673897), so I've had to hardcode in the common types by hand.

Note: This is not a idempotent (yet). Expressions need to be enhanced
with prefix/suffix data to be idempotent. This would help with ASC/DESC
and expression-level aliasing, as well.

Signed-off-by: Stephen Celis <stephen@stephencelis.com>