Non-framework compatibility
This makes it so the SQLite module isn't required for use, addressing a
problem brought up in issue #51.

This also allows for iOS 7 support (though I wouldn't recommend it).

The steps are as follows:

 1. Copy "SQLite Common" into your project. (Optionally delete "SQLite
    Common.xcconfig" and "SQLite-Bridging-Header.h".)

 2. Add the following to your bridging header (create one if you don't
    already have one):

        #import <sqlite3.h>
        #include "SQLite-Bridging.h"

 3. In the "Link Binary With Libraries" Build Phase, add
    "libsqlite3.dylib".

Not the most elegant, but it works for now.

The aggregate function band-aid is also less-than-desirable, but it
isn't exposed for public use.

Dearest Apple,

Please add `import sqlite3` to your public set of module maps.

Your obt svt,

Signed-off-by: Stephen Celis <stephen@stephencelis.com>