Use an embedded module instead of a bridging header
I'd taken advantage of the fact that renaming the umbrella header
disabled the framework bridging header check, which is a bug and could
stop working at any moment.

Instead, let's embed a `sqlite3` framework module that points to the
appropriate system header.

As soon as the system provides an appropriate `sqlite3` module at the
system level, we merely have to delete this internal dependency.

Signed-off-by: Stephen Celis <stephen@stephencelis.com>