Split into standard / standalone subspecs
The standard subspec is the default and behaves
as previously; the standalone subspec adds a
dependency to the sqlite3 pod which downloads and
compiles a version of SQLite instead of using the
system-supplied version.

The version of SQLite used can be further customized
by requiring another sqlite3 subspec from the main
Podfile.