make fromDatatypeValue throw
When using custom types, sometimes decoding can fail, say due to
changes in the type structure. In this case decoding would fail
and the only way to handle it is forcing a crash.

This change allows you to use `try row.get()` instead. Givng you
the chance to handle the mismatch.